Update the test cases
There is one caveat to the test cases, I'm not sure if this is a bug in the sniff, or phpcs or what.

When I put the following code as the test case:

echo (new Components())->render( // Bad
	'accordion',
	Components::props('accordion', $attributes, [
		'accordionContent' => $innerBlockContent
	])
);

Which is clearly an error, and the one that is caught at the correct line,
but the error also gets reported (from the upstream sniff), on the line with the
Components::props call. The same construct is not reported when it's used when
Components::render call is echoed out.

So this odd test case was written with just an array instead.
This should be reported upstream.